% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{note}
\alias{note}
\title{Plays musical notes}
\usage{
note(note, length = 1, octave = 0, volume = default_volume)
}
\arguments{
\item{note}{letter of a musical note (Bb)}

\item{length}{note length (1/2)}

\item{octave}{above and below middle C (-1)}

\item{volume}{volume between 1 - 10}
}
\value{
note object
}
\description{
Plays musical notes
}
\examples{
note("Bb", 1/2, -1, 3)
}
