% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_drm_df.R
\name{play_drm_df}
\alias{play_drm_df}
\title{play_drm_df}
\usage{
play_drm_df(data, drm, duration, volume)
}
\arguments{
\item{data}{A data frame}

\item{drm}{a variable with drm encoding}

\item{duration}{note duration}

\item{volume}{volume duration}
}
\value{
will play notes
}
\description{
play_drm_df
}
