% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_potter_phrase_df.R
\docType{data}
\name{potter_phrase_df}
\alias{potter_phrase_df}
\title{Harry Potter theme data frame}
\format{
\subsection{\code{potter_phrase_df}}{

A data frame with 8 rows:
\describe{
\item{note}{note name}
\item{duration}{duration of note}
\item{octave}{note octave}
...
}
}
}
\source{
\url{https://www.warnerbros.com/movies/harry-potter-complete-8-film-collection}
}
\usage{
potter_phrase_df
}
\description{
Note, duration and  ...
}
\keyword{datasets}
