% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_twinkle_phrase_df.R
\docType{data}
\name{twinkle_phrase_df}
\alias{twinkle_phrase_df}
\title{Harry Potter theme data frame}
\format{
\subsection{\code{twinkle_phrase_df}}{

A data frame with 8 rows:
\describe{
\item{drm}{note name using drm schema}
\item{duration}{duration of note}
...
}
}
}
\source{
\url{https://www.warnerbros.com/movies/harry-potter-complete-8-film-collection}
}
\usage{
twinkle_phrase_df
}
\description{
Note, duration and  ...
}
\keyword{datasets}
