% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\title{Create music}
\usage{
note(note, length = 1, octave = 0, volume = default_volume)
}
\arguments{
\item{note}{name}

\item{length}{in beats}

\item{octave}{from middle C}

\item{volume}{from 1 to 10}
}
\value{
a note object
}
\description{
Create music
}
\examples{
note("A")
}
