% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesterfunctions.R
\name{note}
\alias{note}
\title{Title}
\usage{
note(note, length = 1, octave = 0, volume = default_volume)
}
\arguments{
\item{note}{}

\item{length}{}

\item{octave}{}

\item{volume}{}
}
\description{
Title
}
\examples{
note("A")
}
