% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\title{note}
\usage{
note(
  name = NULL,
  octave = 0,
  drm = NULL,
  duration = 1,
  volume = default_volume
)
}
\arguments{
\item{name}{note name}

\item{octave}{octave}

\item{drm}{drm}

\item{duration}{duration}

\item{volume}{volume}
}
\description{
note
}
\examples{
note("A")
audio::play(c(
note("B", .5, 0),
note("E", .75, 0)))
}
