% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesterfunctions.R
\name{note}
\alias{note}
\title{Create music}
\usage{
note(name, length = 1, octave = 0, volume = default_volume)
}
\arguments{
\item{length}{in beats}

\item{octave}{from middle C}

\item{volume}{from 1 to 10}

\item{note}{name}
}
\value{
a note object
}
\description{
Create music
}
\examples{
note("A")
audio::play(c(
note("B", .5, 0),
note("E", .75, 0),
note("G", .25, 0),
note("F#", .5, 0),
note("E", 1, 0),
note("B", .5, 1),
note("A", 1.5, 1),
note("F#", 1.5, 0)
))
}
