% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_notes_df.R
\name{play_notes_df}
\alias{play_notes_df}
\title{play_notes_df}
\usage{
play_notes_df(data, name, octave, duration, volume)
}
\arguments{
\item{data}{A data frame}

\item{name}{character variable with note names}

\item{octave}{numeric with note octave}

\item{duration}{numeric with note duration}

\item{volume}{numeric with note volume}
}
\description{
play_notes_df
}
